+++ components/display/channel/channel-sidebar.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { Channel, OldChannelType, OldMemberRole, Server } from "@prisma/client";
+import { OldChannelType, OldMemberRole } from "@prisma/client";
 import { redirect } from "next/navigation";
 
 import { ScrollArea } from "@/components/ui/scroll-area";