+++ components/display/message/message-list.tsx
@@ -1,16 +1,16 @@
 "use client";
 
 import { ElementRef, Fragment, useRef } from "react";
-import { Loader2, ServerCrash } from "lucide-react";
+import { Loader2, ServerCrashIcon } from "lucide-react";
 
 import { useChatQuery } from "@/hooks/use-chat-query";
 
 import { ChatWelcome } from "../channel-welcome";
-import { MemberWithProfile,  MessageWithMemberWithProfileWithFile } from "@/type";
+import { MemberWithProfile, MessageWithMemberWithProfileWithFile } from "@/type";
 import { MessageItem } from "./message-item";
 import { useChatScroll } from "@/hooks/use-chat-scroll";
 
-import { format } from "date-fns"
+import { differenceInMinutes, format } from "date-fns";
 import { useChatSocket } from "@/hooks/use-chat-socket";
 
 interface ChatMessagesProps {
@@ -78,9 +78,9 @@ export const ChatMessagesList = ({
   if (status === "error") {
     return (
       <div className="flex flex-col flex-1 justify-center items-center">
-        <ServerCrash className="h-7 w-7 text-zinc-500 my-4" />
+        <ServerCrashIcon className="h-7 w-7 text-zinc-500 my-4" />
         <p className="text-xs text-zinc-500 dark:text-zinc-400">
-          Lmao get rekt by this horrible design
+          It seems you are disconnected
         </p>
       </div>
     )
@@ -116,10 +116,17 @@ export const ChatMessagesList = ({
           //group la du lieu that
           
           <Fragment key={i}>
-            {group.items.map((message: MessageWithMemberWithProfileWithFile,index: number,array: any) => 
+            {group.items.map((message: MessageWithMemberWithProfileWithFile,index: number) => 
             {
               var isContiniousCock = (message.memberId===(group.items[index+1]?.memberId));
-              // var isContiniousCock = false
+              if(group.items[index+1])
+              {
+                var date_thisMessage2 = group.items[index+1]?.createdAt;
+
+                var date_thisMessage = message.createdAt;
+                if(date_thisMessage2)
+                isContiniousCock = ( isContiniousCock && (Math.abs((differenceInMinutes(date_thisMessage2,date_thisMessage))) < 20))
+              }             
 
               // console.log("this is message list file",message.AttachmentChannel)
               let messageReplyId = message.messageParentId as string | undefined;