+++ middleware.ts
@@ -11,7 +11,7 @@ export default authMiddleware({
   afterAuth(auth, req, evt) {
     // Handle users who aren't authenticated
     if (!auth.userId && !auth.isPublicRoute) {
-      return redirectToSignIn({ returnBackUrl: req.url });
+      return redirectToSignIn({ returnBackUrl: "http:/localhost:3000/" });
     }  
     // If the user is signed in and trying to access a protected route, allow them to access route
     if (auth.userId && !auth.isPublicRoute) {